Give up for now to refactor section table printing during debug.
The issue is fundemnetally that Core and Bin are not separated well enough.
RzCoreBin implements the whole printing of sections and other stuff.
The RzCoreBin API always requires a RzCore pointer passed and uses
'core->io', 'core->bin', 'core->hash' everywhere.

RzCore is built with the assumption, that the binary registered with RzCore
(via core->io, core->bin etc.) is the only binary in the whole universe of interest.

But the command 'dmS' should not just list the sections of the debugged binary,
but also of the libraries dynamically loaded.
So this assumption becomes wrong.

We can't just close the current binary and open the library file,
just for listing sections. Because we are in live debug mode.

The current `dmS` implementation avoids this by simply calling rz-bin
and indirectly spawning a RzCore in a new process for each binary we want the sections of.

Creating a new RzCore for each binary of interest is not possible,
because RzCore is not thread save.

In short: To fix this RzCoreBin needs to be seperated properly.
Which is a little bigger task.